requirements.txt
numpy==1.24.3
scikit-learn==1.3.2
opencv-python==4.10.0.84
PyYAML==6.0.1
matplotlib==3.7.5